requirements.txt
openai
newsapi-python
autopep8
python-dotenv